hca/requirements.txt
streamlit==1.32.0
fastapi==0.110.0
uvicorn==0.27.1
anthropic==0.21.3
xgboost==2.0.3
scikit-learn==1.4.2
shap==0.44.1
pandas==2.2.1
openpyxl==3.1.2
pdfplumber==0.10.3
requests==2.31.0
beautifulsoup4==4.12.3
spacy==3.7.4
python-dotenv==1.0.1
sqlalchemy==2.0.28
httpx==0.27.0
plotly==5.19.0
feedparser==6.0.11
aiofiles==23.2.1
python-multipart==0.0.9
